requirements.txt
fastapi[all]
uvicorn
torch
onnx
onnxruntime
torchvision
tensorflow
onnx2keras